requirements.txt
bs4==0.0.1
python-dotenv==0.15.0
requests==2.25.1
